+++ backend/funco-member/src/main/java/com/found_404/funcomember/feignClient/dto/CoinValuation.java
@@ -2,6 +2,6 @@ package com.found_404.funcomember.feignClient.dto;
 
 public record CoinValuation (
 	String ticker,
-	Long price,
+	Double price,
 	Long valuation) {
 }
